==> code/python/analysis-and-visualization/02-player-basis-weights.ipynb <==
import os.path
import numpy as np
import pandas as pd

pd.set_option('display.float_format', lambda x: '%.6f' % x)

## Load data

def load_data(dirname, filename):
    currdir = os.getcwd()
    datadir = os.path.abspath(os.path.join(currdir, dirname))
    return os.path.abspath(os.path.join(datadir, filename))

datadir = '../../../outputs/'
file_nmf_w = 'spatial-basis/nmf_weights.csv'
file_players = 'players-fg/players-summary-retained-sorted.csv'
outdir = '../../../outputs/spatial-basis'

df_nmf_w = pd.read_csv(os.path.join(datadir, file_nmf_w))
df_players = pd.read_csv(os.path.join(datadir, file_players))

bcols = ['b{}'.format(str(x)) for x in range(0, df_nmf_w.shape[1])]
bcols[0] = 'p_id_retained'
bcols
# df_nmf_w.columns = ['p_id', 'b1', 'b2', 'b3', 'b4', 'b5']
df_nmf_w.columns = bcols
df_nmf_w

# df_players.columns = ['p_id', 'player', 'team', 'fg', 'fga', 'points']
df_players

## Normalize

num_basis = df_nmf_w.shape[1]
df_nmf_w_norm = df_nmf_w.copy(deep=True)
df_nmf_w_norm['player'] = df_players.player
df_nmf_w_norm['team'] = df_players.team
for b in range(1, num_basis):
    colname = 'b{}_norm'.format(b)
#     print(df_nmf_w[str(b)])
#     print(df_nmf_w[[str(x) for x in range(0, df_nmf_w.shape[1]-1)]].agg(sum))
#     print(b_norm)
#     print(colname)
    df_nmf_w_norm[colname] = df_nmf_w.apply(lambda row: row['b{}'.format(str(b))] / row[['b{}'.format(str(x)) for x in range(1, num_basis)]].agg(sum), axis=1)

# df_nmf_w_norm.to_csv()

df_nmf_w_norm

df_nmf_w_norm.to_csv('{}/nmf_weights-players-retained.csv'.format(outdir), index=False)

teams = df_nmf_w_norm.team.unique()
for team in teams:
    df_nmf_w_norm[df_nmf_w_norm.team==team].to_csv('{}/nmf_weights-players-retained-{}.csv'.format(outdir, team), index=False)
