==> code/python/analysis-and-visualization/03-nearest-neighbors.ipynb <==
import os.path
import numpy as np
import pandas as pd
from sklearn.neighbors import NearestNeighbors

pd.set_option('display.float_format', lambda x: '%.6f' % x)

## Load data

def load_data(dirname, filename):
    currdir = os.getcwd()
    datadir = os.path.abspath(os.path.join(currdir, dirname))
    return os.path.abspath(os.path.join(datadir, filename))

datadir = '../../../outputs/spatial-basis'
datafile = 'nmf_weights-players-retained.csv'
outdir = '../../../outputs/nearest-neighbor'

df = pd.read_csv(os.path.join(datadir, datafile))

df_norm = df.drop(columns=['b{}'.format(str(x)) for x in range(1, df.shape[1]//2)])
df_norm
# print(range(1, df.shape[1]//2))

## Nearest neighbor

### League-wide

#### Run NN
- convert player basis weights to numpy array
- compute distances and indices of k nearest neighbors for each player

weights_nn_league = df_norm.drop(columns=['p_id_retained', 'player', 'team']).to_numpy()

neighbors = 6
nn_distances = []
nn_indices = []

nn_league = NearestNeighbors(n_neighbors=neighbors, metric='minkowski')
nn_league.fit(weights_nn_league)

for i in range(weights_nn_league.shape[0]):
    nn_league_neighbors = nn_league.kneighbors([weights_nn_league[i]], neighbors)
    nn_distances.append(nn_league_neighbors[0][0])
    nn_indices.append(nn_league_neighbors[1][0])
    
df_nn_distances = pd.DataFrame(nn_distances, columns=['dist_{}'.format(str(x)) for x in range(neighbors)]).drop(columns=['dist_0'])
df_nn_indices = pd.DataFrame(nn_indices, columns=['nn_{}'.format(str(x)) for x in range(neighbors)]).drop(columns=['nn_0'])
df_nn_di = pd.concat([df_nn_indices, df_nn_distances], axis=1)

# df_nn_distances.to_csv(
#     'outputs/data/nn_distances-league-retained.csv', index_label='p_id_retained'
#     )
# df_nn_indices.to_csv(
#     'outputs/data/nn_indices-league-retained.csv', index_label='p_id_retained'
#     )
# df_nn_di.to_csv(
#     'outputs/data/nn_distances-indices-league-retained.csv', index_label='p_id_retained'
#     )

pd.concat([df_nn_indices, df_nn_distances], axis=1)

# df_nn_indices['p_id_retained'] = df_norm.p_id_retained
df_nn_indices['player'] = df_norm.player
df_nn_indices['team'] = df_norm.team

# df_nn_distances['p_id_retained'] = df_norm.p_id_retained
df_nn_distances['player'] = df_norm.player
df_nn_distances['team'] = df_norm.team

df_nn_di['player'] = df_norm.player
df_nn_di['team'] = df_norm.team

df_nn_distances.to_csv(
    '{}/nn_distances-league-retained-with-pt.csv'.format(outdir), index_label='p_id_retained'
    )
df_nn_distances

df_nn_indices.to_csv(
    '{}/nn_indices-league-retained-with-pt.csv'.format(outdir), index_label='p_id_retained'
    )
df_nn_indices

df_nn_di.to_csv(
    '{}/nn_distances-indices-league-retained-with-pt.csv'.format(outdir), index_label='p_id_retained'
    )
df_nn_di

#### Enforce symmetry
- if player k is a neighbor of player j then player j should be a neighbor of player k
- get list of neighbors
    - get original k neighbors
    - for each player j, check if k is a neighbor
        - if yes, add j to k's neighbors 

p_neighbors = []
num_players, num_cols = df_nn_indices.shape
nn_cols = ['nn_{}'.format(str(x)) for x in range (1, num_cols-2)]

for i in range(num_players):
    p_neighbors_list = df_nn_indices[df_nn_indices.player==df_nn_indices['player'][i]][nn_cols].values.tolist()[0]
    p = df_nn_indices[df_nn_indices.player==df_nn_indices['player'][i]].player.values[0]
    t = df_nn_indices[df_nn_indices.team==df_nn_indices['team'][i]].team.values[0]
    for j in range(num_players):
        if i in df_nn_indices[df_nn_indices.player==df_nn_indices['player'][j]][nn_cols].values:
            p_neighbors_list.append(j)
    p_neighbors.append([i, p, t, list(set(p_neighbors_list))])

df_p_neighbors = pd.DataFrame(p_neighbors, columns=['p_id_retained', 'player', 'team', 'neighbors'])

df_p_neighbors.to_csv(
    '{}/nn_players.csv'.format(outdir), index=False
)

df_p_neighbors_stats = df_p_neighbors.copy(deep=True)

# get number of neighbors
df_p_neighbors_stats['num_neighbors'] = df_p_neighbors_stats.apply(lambda row: len(row.neighbors), axis=1)
df_p_neighbors_stats.to_csv(
    '{}/nn_players_2.csv'.format(outdir), index=False
)

df_p_neighbors_stats

### Per-team

#### Initialize variables

team_names = np.unique(df_norm.team.values)
weights_teams = []
b_cols = ['b{}_norm'.format(str(x)) for x in range (1, df_norm.shape[1] - 2)]

# weights_team = [df_nn_distances[df_nn_distances.team==t][dist_cols] for t in team_names]
for t in team_names:
    df_norm[df_norm.team==t].to_csv('{}/nmf_weights-per-team-{}.csv'.format(outdir, t), index=False)
    weights_teams.append(df_norm[df_norm.team==t][b_cols])

#### Run NN
- convert player basis weights to numpy array
- compute distances and indices to all teammates

nn_distances_t = []
nn_indices_t = []

max_p = len(max(weights_teams, key=len))

for t in weights_teams:
    p = len(t)
    nn_team = NearestNeighbors(n_neighbors=p, metric='minkowski')
    nn_team.fit(t)

#     wt = t
#     for i in wt.index:
#         nn_team_neighbors = nn_team.kneighbors([wt[wt.index==i]][0], p)
#         nn_distances_t.append(nn_team_neighbors[0][0])
#         nn_indices_t.append(nn_team_neighbors[1][0])
        
    wt = t.to_numpy()
    for i in range(p):
        nn_team_neighbors = nn_team.kneighbors([wt[i]], p)
        nn_distances_t.append(nn_team_neighbors[0][0])
        nn_indices_t.append(nn_team_neighbors[1][0])

# df_nn_distances_team = pd.DataFrame(nn_distances_t)

df_nn_distances_team = pd.DataFrame(nn_distances_t, columns=['dist_{}'.format(str(x)) for x in range(max_p)]).drop(columns=['dist_0'])
df_nn_indices_team = pd.DataFrame(nn_indices_t, columns=['nn_{}'.format(str(x)) for x in range(max_p)]).drop(columns=['nn_0'])

df_nn_distances_team['p_id_retained'] = df_norm.sort_values(by=["team", "p_id_retained"]).p_id_retained.reset_index(drop=True)
df_nn_distances_team['player'] = df_norm.sort_values(by=["team", "p_id_retained"]).player.reset_index(drop=True)
df_nn_distances_team['team'] = df_norm.sort_values(by=["team", "p_id_retained"]).team.reset_index(drop=True)

df_nn_indices_team['p_id_retained'] = df_norm.sort_values(by=["team", "p_id_retained"]).p_id_retained.reset_index(drop=True)
df_nn_indices_team['player'] = df_norm.sort_values(by=["team", "p_id_retained"]).player.reset_index(drop=True)
df_nn_indices_team['team'] = df_norm.sort_values(by=["team", "p_id_retained"]).team.reset_index(drop=True)

df_nn_distances_team.to_csv(
    '{}/nn_distances-team-retained.csv'.format(outdir), index=False
    )
df_nn_indices_team.to_csv(
    '{}/nn_indices-team-retained.csv'.format(outdir), index=False
    )

df_nn_distances_team
