==> code/python/analysis-and-visualization/05-global-analysis.ipynb <==
%matplotlib inline
import os.path
import numpy as np
import pandas as pd
from scipy import stats
import matplotlib.pyplot as plt
from matplotlib.patches import Circle, Rectangle, Arc

def load_data(dirname, filename):
    currdir = os.getcwd()
    datadir = os.path.abspath(os.path.join(currdir, dirname))
    return os.path.abspath(os.path.join(datadir, filename))

def generate_xy_grids(xrange, yrange):
    x = np.arange(xrange[0], xrange[1], xrange[2])
    y = np.arange(yrange[0], yrange[1], yrange[2])

    return np.meshgrid(x, y)

dirname = '../../../outputs/'
ebppa = 'ebppa/ebppa.csv'
ebppa_wc = 'ebppa/ebppa-with-cell.csv'
outdir = '../../../outputs/global-stats'

df_ebppa = pd.read_csv(load_data(dirname, ebppa_wc))
np_ebppa = df_ebppa.to_numpy()

playersfg = 'players-fg/players-fg-retained.csv'
playersfg_all = 'players-fg/players-fg-all.csv'

df_fg = pd.read_csv(load_data(dirname, playersfg_all))
np_fg = df_fg.to_numpy()

XX,YY = generate_xy_grids((-75,75,5), (-15,125,5))    #xx = array of x-coordinates; yy = array of y-coordinates
GRID_Y, GRID_X = XX.shape                             # gridy = number of rows = cells along sideline (y); gridx = number of cols = cells along baseline (x);

# grid_fga = np.empty(xx.shape)
# grid_fg = np.empty(xx.shape)
# grid_pts = np.empty(xx.shape)
# grid_ppa = np.empty(xx.shape)

def get_player_fg_grid(player):
      
    df_player = df_fg[df_fg.player==player]
    grid_fga = np.empty(XX.shape)
    grid_fg = np.empty(XX.shape)
    grid_pts = np.empty(XX.shape)
    grid_ppa = np.empty(XX.shape)
    
    for y in range(0, GRID_Y):
        for x in range(0, GRID_X):
            fga = df_player[(df_player.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_player.y.isin(range(YY[y][x],YY[y][x] + 5)))]
            fg = df_player[(df_player.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_player.y.isin(range(YY[y][x],YY[y][x] + 5))) & (df_player.made==1)]
            pts = fg.made_points.sum()
            if len(fga) > 0:
                ppa = pts/len(fga)
            else:
                ppa = 0

            grid_fga[y][x] = len(fga)
            grid_fg[y][x] = len(fg)
            grid_pts[y][x] = pts
            grid_ppa[y][x] = ppa
    
    return {'fga': grid_fga, 'fg': grid_fg, 'pts': grid_pts, 'ppa': grid_ppa, 'team': df_player['team'].unique()[0]}
    

def get_team_fg_grid(team):
      
    df_team = df_fg[df_fg.team==team]
    grid_fga = np.empty(XX.shape)
    grid_fg = np.empty(XX.shape)
    grid_pts = np.empty(XX.shape)
    grid_ppa = np.empty(XX.shape)
    
    for y in range(0, GRID_Y):
        for x in range(0, GRID_X):
            fga = df_team[(df_team.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_team.y.isin(range(YY[y][x],YY[y][x] + 5)))]
            fg = df_team[(df_team.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_team.y.isin(range(YY[y][x],YY[y][x] + 5))) & (df_team.made==1)]
            pts = fg.made_points.sum()
            if len(fga) > 0:
                ppa = pts/len(fga)
            else:
                ppa = 0

            grid_fga[y][x] = len(fga)
            grid_fg[y][x] = len(fg)
            grid_pts[y][x] = pts
            grid_ppa[y][x] = ppa
    
    return {'fga': grid_fga, 'fg': grid_fg, 'pts': grid_pts, 'ppa': grid_ppa}

def get_team_opp_fg_grid(team):
      
    df_team = df_fg[df_fg.opponent==team]
    grid_fga = np.empty(XX.shape)
    grid_fg = np.empty(XX.shape)
    grid_pts = np.empty(XX.shape)
    grid_ppa = np.empty(XX.shape)
    
    for y in range(0, GRID_Y):
        for x in range(0, GRID_X):
            fga = df_team[(df_team.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_team.y.isin(range(YY[y][x],YY[y][x] + 5)))]
            fg = df_team[(df_team.x.isin(range(XX[y][x],XX[y][x] + 5))) & (df_team.y.isin(range(YY[y][x],YY[y][x] + 5))) & (df_team.made==1)]
            pts = fg.made_points.sum()
            if len(fga) > 0:
                ppa = pts/len(fga)
            else:
                ppa = 0

            grid_fga[y][x] = len(fga)
            grid_fg[y][x] = len(fg)
            grid_pts[y][x] = pts
            grid_ppa[y][x] = ppa
    
    return {'fga': grid_fga, 'fg': grid_fg, 'pts': grid_pts, 'ppa': grid_ppa}

df_a = df_fg[df_fg.player=='M. Lee']
df_a['team'].unique()[0]

players = df_fg['player'].unique()
teams = df_fg['team'].unique()

pl = 'M. Lee'
grid_pl = get_player_fg_grid(pl)
ppak = np.sum(grid_pl['pts'])/np.sum(grid_pl['fga'])
elptsk = np.multiply(grid_pl['fga'],np_ebppa)
eppak = np.sum(elptsk)/np.sum(grid_pl['fga'])
print('Player: {}'.format(pl))
print('PPA: {}'.format(ppak.round(3)))
print('EPPA: {}'.format(eppak.round(3)))
print('SScE: {}'.format((ppak - eppak.round(3)).round(3)))

stats_players = []

for pl in players:
    grid_pl = get_player_fg_grid(pl)
    team = grid_pl['team']
    fga = np.sum(grid_pl['fga'])
    fg = np.sum(grid_pl['fg'])
    pts = np.sum(grid_pl['pts']).round(0)
    ppa = (pts/fga)
#     ppa = (np.sum(grid_pl['pts'])/np.sum(grid_pl['fga'])).round(3)
    elptsk = (np.multiply(grid_pl['fga'], np_ebppa))
    elpts = np.sum(elptsk)
    eppa = (elpts/np.sum(grid_pl['fga']))
    ssce = (ppa - eppa)
    prla = (pts - elpts)
    
    stats_players.append([pl, team, ssce, prla, eppa, elpts, ppa, pts, fga, fg])

# print(stats_players)

df_stats_players = pd.DataFrame(stats_players, columns=['player', 'team', 'ssce', 'prla', 'eppa', 'elpts', 'ppa', 'pts', 'fga', 'fg'])

df_stats_players.to_csv("{}/player-analysis-global-all.csv".format(outdir), index=False)

stats_teams = []

for pl in teams:
    grid_pl = get_team_fg_grid(pl)
    fga = np.sum(grid_pl['fga'])
    fg = np.sum(grid_pl['fg'])
    pts = np.sum(grid_pl['pts']).round(0)
    ppa = (pts/fga)
#     ppa = (np.sum(grid_pl['pts'])/np.sum(grid_pl['fga'])).round(3)
    elptsk = (np.multiply(grid_pl['fga'], np_ebppa))
    elpts = np.sum(elptsk)
    eppa = (elpts/np.sum(grid_pl['fga']))
    ssce = (ppa - eppa)
    prla = (pts - elpts)
    
    stats_teams.append([pl, ssce, prla, eppa, elpts, ppa, pts, fga, fg])

df_stats_teams = pd.DataFrame(stats_teams, columns=['team', 'ssce', 'prla', 'eppa', 'elpts', 'ppa', 'pts', 'fga', 'fg'])

df_stats_teams.to_csv("{}/team-analysis-global-all.csv".format(outdir), index=False)

stats_teams_opp = []

for pl in teams:
    grid_pl = get_team_opp_fg_grid(pl)
    fga = np.sum(grid_pl['fga'])
    fg = np.sum(grid_pl['fg'])
    pts = np.sum(grid_pl['pts']).round(0)
    ppa = (pts/fga)
#     ppa = (np.sum(grid_pl['pts'])/np.sum(grid_pl['fga'])).round(3)
    elptsk = (np.multiply(grid_pl['fga'], np_ebppa))
    elpts = np.sum(elptsk)
    eppa = (elpts/np.sum(grid_pl['fga']))
    ssce = (ppa - eppa)
    prla = (pts - elpts)
    
    stats_teams_opp.append([pl, ssce, prla, eppa, elpts, ppa, pts, fga, fg])


df_stats_teams_opp = pd.DataFrame(stats_teams_opp, columns=['team', 'opp_ssce', 'opp_prla', 'opp_eppa', 'opp_elpts', 'opp_ppa', 'opp_pts', 'opp_fga', 'opp_fg'])

df_stats_teams_opp.to_csv("{}/team-analysis-opp-global-all.csv".format(outdir), index=False)

df_stats = pd.read_csv("{}/player-analysis-global-complete-retained.csv".format(outdir), sep=',', header=0)

x_ssce = df_stats.ssce.to_numpy()
x_prla = df_stats.prla.to_numpy()
x_efg = df_stats.efg.to_numpy()
x_ppa = df_stats.ppa.to_numpy()
x_fga = df_stats.fga.to_numpy()
x_pts = df_stats.pts.to_numpy()

print('ssce, efg:', stats.pearsonr(x_ssce, x_efg))
print('prla, efg:', stats.pearsonr(x_prla, x_efg))
# print('ssce, ppa:', stats.pearsonr(x_ssce, x_ppa))
print('ssce, fga:', stats.pearsonr(x_ssce, x_fga))
print('prla, fga:', stats.pearsonr(x_prla, x_fga))
print('ssce, pts:', stats.pearsonr(x_ssce, x_pts))
print('prla, pts:', stats.pearsonr(x_prla, x_pts))

print('ssce, pts:', stats.spearmanr(x_ssce, x_pts))
print('prla, pts:', stats.spearmanr(x_prla, x_pts))
print('ssce, fga:', stats.spearmanr(x_ssce, x_fga))
print('ssce, fga:', stats.kendalltau(x_ssce, x_fga))

# np.corrcoef(x_ssce, x_efg)

print(x_ssce)
print(x_efg)

stats.pearsonr(x_ssce, x_efg)


df_statscomp = pd.read_csv("{}/local-stats/players-statscomp.csv".format(dirname), sep=',', header=0)

x_spread = df_statscomp.fga_cells.to_numpy()
x_fga = df_statscomp.fga.to_numpy()
x_terng = df_statscomp.terange.to_numpy()
x_erng = df_statscomp.erange.to_numpy()
x_perng = df_statscomp.perange.to_numpy()
x_nerng = df_statscomp.nerange.to_numpy()


print('spread, fga:', stats.pearsonr(x_fga, x_spread))
print('erng, spread:', stats.pearsonr(x_spread, x_erng))
print('terng, spread:', stats.pearsonr(x_spread, x_terng))
print('perng, spread:', stats.pearsonr(x_spread, x_perng))
print('nerng, spread:', stats.pearsonr(x_spread, x_nerng))
print('perng, terng:', stats.pearsonr(x_terng, x_perng))

print('erng, fga:', stats.pearsonr(x_fga, x_erng))
print('terng, fga:', stats.pearsonr(x_fga, x_terng))
print('nerng, fga:', stats.pearsonr(x_fga, x_nerng))
print('perng, fga:', stats.pearsonr(x_fga, x_perng))


df_pythag = pd.read_csv("{}/team-analysis/pythag.csv".format(dirname), sep=',', header=0)

x_pwp = df_pythag.pwp.to_numpy()
x_rank_p = df_pythag.rank_p.to_numpy()
x_rank_s = df_pythag.rank_s.to_numpy()
x_rank_ps = df_pythag.rank_ps.to_numpy()

print('rank_p, rank_s:', stats.pearsonr(x_rank_p, x_rank_s))
print('rank_p, rank_s:', stats.spearmanr(x_rank_p, x_rank_s))
print('rank_p, rank_s:', stats.kendalltau(x_rank_p, x_rank_s))
print('pwp, rank_s:', stats.pearsonr(x_pwp, x_rank_s))
print('pwp, rank_s:', stats.spearmanr(x_pwp, x_rank_s))
print('pwp, rank_s:', stats.kendalltau(x_pwp, x_rank_s))
print('rank_ps, rank_s:', stats.pearsonr(x_rank_ps, x_rank_s))
print('rank_ps, rank_s:', stats.spearmanr(x_rank_ps, x_rank_s))
print('rank_ps, rank_s:', stats.kendalltau(x_rank_ps, x_rank_s))
