==> code/python/analysis-and-visualization/07-metric-analysis.ipynb <==
%matplotlib inline
import pickle
import os.path
import numpy as np
from scipy import stats
from statsmodels.stats import weightstats
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib.patches import Circle, Rectangle, Arc

def load_data(dirname, filename):
    currdir = os.getcwd()
    datadir = os.path.abspath(os.path.join(currdir, dirname))
    return os.path.abspath(os.path.join(datadir, filename))

def generate_xy_grids(xrange, yrange):
    x = np.arange(xrange[0], xrange[1], xrange[2])
    y = np.arange(yrange[0], yrange[1], yrange[2])

    return np.meshgrid(x, y)

def draw_court(ax=None, color='black', lw=2, outer_lines=False):
    # If an axes object isn't provided to plot onto, just get current one
    if ax is None:
        ax = plt.gca()

    # Create the various parts of an NBA basketball court

    # Create the basketball hoop
    # Diameter of a hoop is 46cm so it has a radius of 23cm, which is a value
    # 2.3 in our coordinate system
    hoop = Circle((0, 0), radius=2.3, linewidth=lw, color=color, fill=False)

    # Create backboard
    backboard = Rectangle((-9, -2.3), 18, -0.1, linewidth=lw, color=color)

    # The paint
    # Create the outer box 0f the paint, width=4.9m, height=5.8m
    outer_box = Rectangle((-24.45, -14.3), 49, 58, linewidth=lw, color=color,
                          fill=False)
    # Create the inner box of the paint, widt=3.6ft, height=5.8ft
    inner_box = Rectangle((-18, -14.3), 36, 58, linewidth=lw, color=color,
                          fill=False)

    # Create free throw top arc
    top_free_throw = Arc((0, 43.7), 36, 36, theta1=0, theta2=180,
                         linewidth=lw, color=color, fill=False)
    # Create free throw bottom arc
    bottom_free_throw = Arc((0, 43.7), 36, 36, theta1=180, theta2=0,
                            linewidth=lw, color=color, linestyle='dashed')
    # Restricted Zone, it is an arc with 4ft radius from center of the hoop
    restricted = Arc((0, 0), 25, 25, theta1=0, theta2=180, linewidth=lw,
                     color=color)

    # Three point line
    # Create the side 3pt lines, they are 2.68m long before they begin to arc
    corner_three_a = Rectangle((-66, -14.3), 0, 28.8, linewidth=lw,
                               color=color)
    corner_three_b = Rectangle((66, -14.3), 0, 28.8, linewidth=lw, color=color)
    # 3pt arc - center of arc will be the hoop, arc is 6.75m away from hoop
    # I just played around with the theta values until they lined up with the 
    # threes
    three_arc = Arc((0, 0), 135, 135, theta1=12, theta2=168, linewidth=lw,
                    color=color)

    # Center Court
    center_outer_arc = Arc((0, 125.7), 36, 36, theta1=180, theta2=0,
                           linewidth=lw, color=color)
    center_inner_arc = Arc((0, 125.7), 12, 12, theta1=180, theta2=0,
                           linewidth=lw, color=color)

    # List of the court elements to be plotted onto the axes
    court_elements = [
        hoop,
        backboard,
        outer_box,
        inner_box,
        top_free_throw,
        bottom_free_throw,
        restricted,
        corner_three_a,
        corner_three_b,
        three_arc,
        center_outer_arc,
        center_inner_arc
    ]

    if outer_lines:
        # Draw the half court line, baseline and side out bound lines
        outer_lines = Rectangle((-75, -14.3), 150, 140, linewidth=lw,
                                color=color, fill=False)
        court_elements.append(outer_lines)

    # Add the court elements onto the axes
    for element in court_elements:
        ax.add_patch(element)

    return ax

dirname = '../../../outputs/'
ebppa = 'ebppa/ebppa.csv'
outdir = '../../../outputs/local-stats'

playersfg = 'players-fg/players-fg-retained.csv'

df_fg = pd.read_csv(load_data(dirname, playersfg))
players = df_fg['player'].unique()

filehandler = open('{}/players-local-stats.obj'.format(outdir), 'rb') 
df_stats_players = pickle.load(filehandler)

filehandler_teams = open('{}/teams-local-stats.obj'.format(outdir), 'rb') 
df_stats_teams = pickle.load(filehandler_teams)

filehandler_teams_opp = open('{}/teams-opp-local-stats.obj'.format(outdir), 'rb') 
df_stats_teams_opp = pickle.load(filehandler_teams_opp)

pl = []
for p in players:
    df_x = df_stats_players[df_stats_players.player==p]
    team = df_x.team.values[0]
    
    df_x_ssce = df_x.grid_ssce
    df_x_fga = df_x.grid_fga

    x_ssce = df_x_ssce.to_numpy()[0]
    x_fga = df_x_fga.to_numpy()[0]

    num_fga_cells = (x_fga > 0).sum()   #N

    fga = np.multiply(x_fga,(x_fga > 0)).sum()
    wssce = np.multiply(x_ssce, x_fga).sum()
    sse = wssce/fga  # weighted mean SSE
    wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
    wss_den = fga/num_fga_cells
    wss = wss_num/wss_den
    s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
    stdev = np.sqrt(s2) # weighted stdev
    
    sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
#     aaa = stdev/np.sqrt(num_fga_cells)
    tstat = sse/sterr  # t-statistic
    a = 2.048
    if num_fga_cells >= 120:
        a = 1.980
    elif num_fga_cells in range(60,120):
        a = 2.000
    elif num_fga_cells in range(30,60):
        a = 2.042
        
    ci = a * sterr
    upper = sse + ci
    lower = sse - ci
    
    pl.append([p, team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, a, tstat < a, wss])

pd.DataFrame(pl, columns=['player', 'team', 'fga', 'fga_cells', 'w_ssce', 'upper', 'lower', 'ci95', 'w_var', 'stdev', 'sterr', 'tstat', 'tval', 'ftr', 'wss']).to_csv('{}/players-ssce-95ci.csv'.format(outdir), index=False)

# p = 'F. Jaboneta'
# df_x = df_stats_players[df_stats_players.player==p]

# df_x_ssce = df_x.grid_ssce
# df_x_fga = df_x.grid_fga

# x_ssce = df_x_ssce.to_numpy()[0]
# x_fga = df_x_fga.to_numpy()[0]

# num_fga_cells = (x_fga > 0).sum()

# fga = np.multiply(x_fga,(x_fga > 0)).sum()
# wssce = np.multiply(x_ssce, x_fga).sum()
# sse = wssce/fga
# wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
# wss_den = fga/num_fga_cells
# wss = wss_num/wss_den
# s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)
# stdev = np.sqrt(s2)

# aaa = stdev/np.sqrt(num_fga_cells)
# a = 2.048
# if num_fga_cells >= 120:
#     a = 1.980
# elif num_fga_cells in range(60,120):
#     a = 2.000
# elif num_fga_cells in range(30,60):
#     a = 2.042
# # a = 1.701
# # if num_fga_cells >= 120:
# #     a = 1.658
# # elif num_fga_cells in range(60,120):
# #     a = 1.671
# # elif num_fga_cells in range(30,60):
# #     a = 1.697
    
# ci = a * aaa

# print('fga:', fga)
# print('N:', num_fga_cells)
# print('ssce:', round(sse,5))
# print('wss:', wss)
# print('s2:', s2)
# print('stdev:', stdev)
# print('a:', a)
# print('aaa:', aaa)
# print('ci:', ci)
# print('upper:', sse + ci)
# print('lower:', sse - ci)


p = 'J. Ahanmisi'
df_x = df_stats_players[df_stats_players.player==p]
team = df_x.team.values[0]

df_x_pts = df_x.grid_pts
df_x_elpts = df_x.grid_elpts

x_pts = df_x_pts.to_numpy()[0]
x_elpts = df_x_elpts.to_numpy()[0]

num_fga_cells = (x_elpts > 0).sum()   #N

lprla = np.subtract(x_pts, x_elpts).sum()
# lprla_mean = lprla/num_fga_cells
# wssce = np.multiply(x_ssce, x_fga).sum()
# sse = wssce/fga
# wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
# wss_den = fga/num_fga_cells
# wss = wss_num/wss_den
# s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)
# stdev = np.sqrt(s2)

# aaa = stdev/np.sqrt(num_fga_cells)
# a = 2.048
# if num_fga_cells >= 120:
#     a = 1.980
# elif num_fga_cells in range(60,120):
#     a = 2.000
# elif num_fga_cells in range(30,60):
#     a = 2.042
# # a = 1.701
# # if num_fga_cells >= 120:
# #     a = 1.658
# # elif num_fga_cells in range(60,120):
# #     a = 1.671
# # elif num_fga_cells in range(30,60):
# #     a = 1.697
    
# ci = a * aaa

# print('fga:', fga)
print('N:', num_fga_cells)
print('lprla:', lprla.sum())
# print('wss:', wss)
# print('s2:', s2)
# print('stdev:', stdev)
# print('a:', a)
# print('aaa:', aaa)
# print('ci:', ci)
# print('upper:', sse + ci)
# print('lower:', sse - ci)


# pl2 = []
# for p in players:
#     df_x = df_stats_players[df_stats_players.player==p]
#     team = df_x.team.values[0]
    
#     df_x_pts = df_x.grid_pts
#     df_x_elpts = df_x.grid_elpts

#     x_pts = df_x_pts.to_numpy()[0]
#     x_elpts = df_x_elpts.to_numpy()[0]

#     num_fga_cells = (x_pts > 0).sum()   #N

#     lprla = np.subtract(x_pts, x_elpts)
#     #fga = np.multiply(x_fga,(x_fga > 0)).sum()
#     #wssce = np.multiply(x_ssce, x_fga).sum()
#     sse = wssce/fga  # weighted mean SSE
#     wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
#     wss_den = fga/num_fga_cells
#     wss = wss_num/wss_den
#     s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
#     stdev = np.sqrt(s2) # weighted stdev
    
#     sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
# #     aaa = stdev/np.sqrt(num_fga_cells)
#     tstat = sse/sterr  # t-statistic
#     a = 2.048
#     if num_fga_cells >= 120:
#         a = 1.980
#     elif num_fga_cells in range(60,120):
#         a = 2.000
#     elif num_fga_cells in range(30,60):
#         a = 2.042
        
#     ci = a * sterr
#     upper = sse + ci
#     lower = sse - ci
    
#     pl2.append([p, team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, a, tstat < a, wss])

pl = []
for p in players:
    df_x = df_stats_players[df_stats_players.player==p]
    team = df_x.team.values[0]
    
    df_x_ssce = df_x.grid_ssce
    df_x_fga = df_x.grid_fga

    x_ssce = df_x_ssce.to_numpy()[0]
    x_fga = df_x_fga.to_numpy()[0]

    num_fga_cells = (x_fga > 0).sum()   #N

    fga = np.multiply(x_fga,(x_fga > 0)).sum()
    wssce = np.multiply(x_ssce, x_fga).sum()
    sse = wssce/fga  # weighted mean SSE
    wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
    wss_den = fga/num_fga_cells
    wss = wss_num/wss_den
    s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
    stdev = np.sqrt(s2) # weighted stdev
    
    sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
#     aaa = stdev/np.sqrt(num_fga_cells)
    tstat = sse/sterr  # t-statistic
    tcrit = 2.048
    if num_fga_cells == 29:
        tcrit = 2.045
    elif num_fga_cells == 30:
        tcrit = 2.042
    elif num_fga_cells in range (31,40):
        tcrit = 2.021
    elif num_fga_cells in range (40,50):
        tcrit = 2.015
    elif num_fga_cells in range (50,60):
        tcrit = 2.009
    elif num_fga_cells in range (60,70):
        tcrit = 2.000
    elif num_fga_cells in range (70,80):
        tcrit = 1.994
    elif num_fga_cells in range (80,90):
        tcrit = 1.990
    elif num_fga_cells in range (90,100):
        tcrit = 1.987
    elif num_fga_cells in range (100,120):
        tcrit = 1.984
    elif num_fga_cells in range (120,150):
        tcrit = 1.980
    elif num_fga_cells in range (150,200):
        tcrit = 1.976
    elif num_fga_cells >= 200:
        tcrit = 1.972
        
    ci = tcrit * sterr
    upper = sse + ci
    lower = sse - ci
    
    erange = (x_ssce > 0).sum()
    nerange = erange - (np.multiply(x_fga, x_ssce) < 0).sum()
    perange = round(100*erange/num_fga_cells, 3)
    terange = round(100*erange/587, 3)
    
    pl.append([p, team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, tcrit, tstat < tcrit, wss, erange, nerange, perange, terange])

pd.DataFrame(pl, columns=['player', 'team', 'fga', 'fga_cells', 'w_ssce', 'upper', 'lower', 'ci95', 'w_var', 'stdev', 'sterr', 'tstat', 'tval', 'ftr', 'wss', 'erange', 'nerange', 'perange', 'terange']).to_csv('{}/players-statscomp.csv'.format(outdir), index=False)

df_plstat = pd.DataFrame(pl, columns=['player', 'team', 'fga', 'fga_cells', 'w_ssce', 'upper', 'lower', 'ci95', 'w_var', 'stdev', 'sterr', 'tstat', 'tval', 'ftr', 'wss', 'erange', 'nerange', 'perange', 'terange'])

df_plstat

df_stats = pd.read_csv("{}/global-stats/player-global-all-stats-retained.csv".format(dirname), sep=',', header=0)

p1n = 'J. Ahanmisi'
p1 = df_stats[df_plstat.player==p1n]

def get_sp_df(p1, p2):
    n1 = p1.fga_cells.values[0]
    n2 = p2.fga_cells.values[0]
    s1 = p1.stdev.values[0]
    s2 = p2.stdev.values[0]
#     print(n1, n2, s1, s2)
    nume = ((n1 - 1)*(s1*s1)) + ((n2 - 1)*(s2*s2))
    deno = n1 + n2 - 2
    sp = np.sqrt(nume/deno)
    
#     print(sp)
    return (sp, deno)

def get_se(p1, p2):
    sp = get_sp_df(p1, p2)[0]
    n1 = p1.fga_cells.values[0]
    n2 = p2.fga_cells.values[0]
    se = sp * np.sqrt((1.0/n1)+(1.0/n2))
    return se

# def get_t(p1, p2):
#     x1 = p1.w_ssce.values[0]
#     x2 = p2.w_ssce.values[0]
#     sp, df =  get_sp_df(p1, p2)
#     se = get_se(p1, p2)
    
#     t = (x1 - x2)/se
    
#     return(t, df, sp, se)


def get_t(p1, p2, tc):
    x1 = p1[tc].values[0]
    x2 = p2[tc].values[0]
    sp, df =  get_sp_df(p1, p2)
    se = get_se(p1, p2)
    
    t = (x1 - x2)/se
    
    return(t, df, sp, se)

# print(get_t(ahan, df_plstat[df_plstat.player=='A. Caracut']))

px = np.in1d(players, [p1n])
players0 = players[~px]

statcomp = 'prla'
pn = []
for p in players0:
    p2 = df_stats[df_plstat.player==p]
    team = p2.team.values[0]
    t, df, sd, se = get_t(p1, p2, statcomp)
    pval = stats.t.sf(abs(t), df=df)
    pn.append([p, team, pval, pval*2, t, df, sd, se])

pd.DataFrame(pn, columns=['player', 'team', 'p-value 1t', 'p-value 2t', 'unpaired-t', 'dof', 'sp', 'se']).to_csv('{}/players-comp-{}-{}.csv'.format(outdir, p1n, statcomp), index=False)

df_ebppa = pd.read_csv(load_data(dirname, ebppa))
np_ebppa = df_ebppa.to_numpy()

pwst = []
for p in players:
    df_x = df_stats_players[df_stats_players.player==p]
    team = df_x.team.values[0]

    df_x_ssce = df_x.grid_ssce
    df_x_fga = df_x.grid_fga
    df_x_prla = df_x.grid_prla

    x_ssce = df_x_ssce.to_numpy()[0]
    x_fga = df_x_fga.to_numpy()[0]
    x_prla = df_x_prla.to_numpy()[0]
    x_ppa = df_x.grid_lppa.to_numpy()[0]
    # x_eppa = df_x.grid_ppa.to_numpy()[0]


    zeroattempts = np.in1d(x_fga, [0])
    # players0 = players[~px]

    # print(zeroattempts)
    # print(x_fga)

    z_w = x_fga[~zeroattempts.reshape(x_fga.shape)]  # weights
#     z_x1 = x_ppa[~zeroattempts.reshape(x_fga.shape)]  #x, 
#     z_x2 = np_ebppa[~zeroattempts.reshape(x_fga.shape)]  #x2
    z_x1 = df_x.grid_pts.to_numpy()[0][~zeroattempts.reshape(x_fga.shape)]  #x1
    z_x2 = df_x.grid_elpts.to_numpy()[0][~zeroattempts.reshape(x_fga.shape)]  #x1, 

    (px, (tlow, plow, doflow), (tupp, pupp, dofupp)) = weightstats.ttost_paired(z_x1, z_x2, 0, 0, weights=z_w)
    
    pwst.append([p, team, tlow, plow, tupp, pupp, doflow])
    
pd.DataFrame(pwst, columns=['player', 'team', 'tlow', 'plow', 'tupp', 'pupp', 'dof']).to_csv('{}/players-ttost-prla.csv'.format(outdir), index=False)
# np.subtract(np.multiply(z_w, z_x1),np.multiply(z_w, z_x2)).sum()/z_w.sum()


# pl = []
# for p in players:
#     df_x = df_stats_players[df_stats_players.player==p]
#     team = df_x.team.values[0]
    
#     df_x_ssce = df_x.grid_ssce
#     df_x_
#     df_x_fga = df_x.grid_fga
#     df_x_prla = df_x.grid_prla

#     x_ssce = df_x_ssce.to_numpy()[0]
#     x_fga = df_x_fga.to_numpy()[0]
#     x_prla = df_x_prla.to_numpy()[0]

#     num_fga_cells = (x_fga > 0).sum()   #N
    

#     fga = np.multiply(x_fga,(x_fga > 0)).sum()
#     wssce = np.multiply(x_ssce, x_fga).sum()
#     sse = wssce/fga  # weighted mean SSE
#     wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
#     wss_den = fga/num_fga_cells
#     wss = wss_num/wss_den
#     s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
#     stdev = np.sqrt(s2) # weighted stdev
    
#     sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
# #     aaa = stdev/np.sqrt(num_fga_cells)
#     tstat = sse/sterr  # t-statistic
#     a = 2.048
#     if num_fga_cells >= 120:
#         a = 1.980
#     elif num_fga_cells in range(60,120):
#         a = 2.000
#     elif num_fga_cells in range(30,60):
#         a = 2.042
        
#     ci = a * sterr
#     upper = sse + ci
#     lower = sse - ci
    
#     pl.append([p, team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, a, tstat < a, wss])

teams = df_stats_teams.team.unique()

tms = []
for p in teams:
    df_x = df_stats_teams[df_stats_teams.team==p]
    team = p
    
    df_x_ssce = df_x.grid_ssce
    df_x_fga = df_x.grid_fga

    x_ssce = df_x_ssce.to_numpy()[0]
    x_fga = df_x_fga.to_numpy()[0]

    num_fga_cells = (x_fga > 0).sum()   #N

    fga = np.multiply(x_fga,(x_fga > 0)).sum()
    wssce = np.multiply(x_ssce, x_fga).sum()
    sse = wssce/fga  # weighted mean SSE
    wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
    wss_den = fga/num_fga_cells
    wss = wss_num/wss_den
    s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
    stdev = np.sqrt(s2) # weighted stdev
    
    sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
#     aaa = stdev/np.sqrt(num_fga_cells)
    tstat = sse/sterr  # t-statistic
    tcrit = 2.048
    if num_fga_cells == 29:
        tcrit = 2.045
    elif num_fga_cells == 30:
        tcrit = 2.042
    elif num_fga_cells in range (31,40):
        tcrit = 2.021
    elif num_fga_cells in range (40,50):
        tcrit = 2.015
    elif num_fga_cells in range (50,60):
        tcrit = 2.009
    elif num_fga_cells in range (60,70):
        tcrit = 2.000
    elif num_fga_cells in range (70,80):
        tcrit = 1.994
    elif num_fga_cells in range (80,90):
        tcrit = 1.990
    elif num_fga_cells in range (90,100):
        tcrit = 1.987
    elif num_fga_cells in range (100,120):
        tcrit = 1.984
    elif num_fga_cells in range (120,150):
        tcrit = 1.980
    elif num_fga_cells in range (150,200):
        tcrit = 1.976
    elif num_fga_cells >= 200:
        tcrit = 1.972
        
    ci = tcrit * sterr
    upper = sse + ci
    lower = sse - ci
    
    erange = (x_ssce > 0).sum()
    nerange = erange - (np.multiply(x_fga, x_ssce) < 0).sum()
    perange = round(100*erange/num_fga_cells, 3)
    terange = round(100*erange/587, 3)
    
    tms.append([team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, tcrit, abs(tstat) < tcrit, wss, erange, nerange, perange, terange])

pd.DataFrame(tms, columns=['team', 'fga', 'fga_cells', 'w_ssce', 'upper', 'lower', 'ci95', 'w_var', 'stdev', 'sterr', 'tstat', 'tval', 'ftr', 'wss', 'erange', 'nerange', 'perange', 'terange']).to_csv('{}/teams-statscomp.csv'.format(outdir), index=False)

teams = df_stats_teams_opp.team.unique()

tms = []
for p in teams:
    df_x = df_stats_teams_opp[df_stats_teams_opp.team==p]
    team = p
    
    df_x_ssce = df_x.grid_ssce
    df_x_fga = df_x.grid_fga

    x_ssce = df_x_ssce.to_numpy()[0]
    x_fga = df_x_fga.to_numpy()[0]

    num_fga_cells = (x_fga > 0).sum()   #N

    fga = np.multiply(x_fga,(x_fga > 0)).sum()
    wssce = np.multiply(x_ssce, x_fga).sum()
    sse = wssce/fga  # weighted mean SSE
    wss_num = np.multiply(np.multiply(x_ssce, x_ssce), x_fga).sum()
    wss_den = fga/num_fga_cells
    wss = wss_num/wss_den
    s2 = (wss - (num_fga_cells * sse * sse))/(num_fga_cells - 1)  # weighted variance
    stdev = np.sqrt(s2) # weighted stdev
    
    sterr = stdev/np.sqrt(num_fga_cells)  # weighted standard error
#     aaa = stdev/np.sqrt(num_fga_cells)
    tstat = sse/sterr  # t-statistic
    tcrit = 2.048
    if num_fga_cells == 29:
        tcrit = 2.045
    elif num_fga_cells == 30:
        tcrit = 2.042
    elif num_fga_cells in range (31,40):
        tcrit = 2.021
    elif num_fga_cells in range (40,50):
        tcrit = 2.015
    elif num_fga_cells in range (50,60):
        tcrit = 2.009
    elif num_fga_cells in range (60,70):
        tcrit = 2.000
    elif num_fga_cells in range (70,80):
        tcrit = 1.994
    elif num_fga_cells in range (80,90):
        tcrit = 1.990
    elif num_fga_cells in range (90,100):
        tcrit = 1.987
    elif num_fga_cells in range (100,120):
        tcrit = 1.984
    elif num_fga_cells in range (120,150):
        tcrit = 1.980
    elif num_fga_cells in range (150,200):
        tcrit = 1.976
    elif num_fga_cells >= 200:
        tcrit = 1.972
        
    ci = tcrit * sterr
    upper = sse + ci
    lower = sse - ci
    
    erange = (x_ssce > 0).sum()
    nerange = erange - (np.multiply(x_fga, x_ssce) < 0).sum()
    perange = round(100*erange/num_fga_cells, 3)
    terange = round(100*erange/587, 3)
    
    tms.append([team, fga, num_fga_cells, sse, upper, lower, ci, s2, stdev, sterr, tstat, tcrit, abs(tstat) < tcrit, wss, erange, nerange, perange, terange])

pd.DataFrame(tms, columns=['team', 'fga', 'fga_cells', 'w_ssce', 'upper', 'lower', 'ci95', 'w_var', 'stdev', 'sterr', 'tstat', 'tval', 'ftr', 'wss', 'erange', 'nerange', 'perange', 'terange']).to_csv('{}/teams-opp-statscomp.csv'.format(outdir), index=False)

df_stats_teams2 = pd.read_csv("{}/local-stats/teams-opp-statscomp.csv".format(dirname), sep=',', header=0)


def get_sp_df(p1, p2):
    n1 = p1.fga_cells.values[0]
    n2 = p2.fga_cells.values[0]
    s1 = p1.stdev.values[0]
    s2 = p2.stdev.values[0]
#     print(n1, n2, s1, s2)
    nume = ((n1 - 1)*(s1*s1)) + ((n2 - 1)*(s2*s2))
    deno = n1 + n2 - 2
    sp = np.sqrt(nume/deno)
    
#     print(sp)
    return (sp, deno)

def get_se(p1, p2):
    sp = get_sp_df(p1, p2)[0]
    n1 = p1.fga_cells.values[0]
    n2 = p2.fga_cells.values[0]
    se = sp * np.sqrt((1.0/n1)+(1.0/n2))
    return se

# def get_t(p1, p2):
#     x1 = p1.w_ssce.values[0]
#     x2 = p2.w_ssce.values[0]
#     sp, df =  get_sp_df(p1, p2)
#     se = get_se(p1, p2)
    
#     t = (x1 - x2)/se
    
#     return(t, df, sp, se)


def get_t(p1, p2, tc):
    x1 = p1[tc].values[0]
    x2 = p2[tc].values[0]
    sp, df =  get_sp_df(p1, p2)
    se = get_se(p1, p2)
    
    t = (x1 - x2)/se
    
    return(t, df, sp, se)

# print(get_t(ahan, df_plstat[df_plstat.player=='A. Caracut']))

p1n = 'ADMU'
p1 = df_stats_teams2[df_stats_teams2.team==p1n]

px = np.in1d(teams, [p1n])
teams0 = teams[~px]

statcomp = 'w_ssce'
tn = []
for p in teams0:
    p2 = df_stats_teams2[df_stats_teams2.team==p]
    t, df, sd, se = get_t(p1, p2, statcomp)
    pval = stats.t.sf(abs(t), df=df)
    tn.append([p, pval, pval*2, t, df, sd, se])

pd.DataFrame(tn, columns=['team', 'p-value 1t', 'p-value 2t', 'unpaired-t', 'dof', 'sp', 'se']).to_csv('{}/teams-opp-comp-{}-{}.csv'.format(outdir, p1n, statcomp), index=False)
